+++ src/client/app/index.ts
@@ -2,7 +2,6 @@ import {
   App,
   createApp as createClientApp,
   createSSRApp,
-  defineAsyncComponent,
   h,
   onMounted,
   watch
@@ -46,8 +45,6 @@ const VitePressApp = {
 export function createApp() {
   const router = newRouter()
 
-  handleHMR(router)
-
   const app = newApp()
 
   app.provide(RouterSymbol, router)
@@ -58,12 +55,6 @@ export function createApp() {
   // install global components
   app.component('Content', Content)
   app.component('ClientOnly', ClientOnly)
-  app.component(
-    'Debug',
-    import.meta.env.PROD
-      ? () => null
-      : defineAsyncComponent(() => import('./components/Debug.vue'))
-  )
 
   // expose $frontmatter
   Object.defineProperty(app.config.globalProperties, '$frontmatter', {
@@ -78,6 +69,13 @@ export function createApp() {
       router,
       siteData: siteDataRef
     })
+  }
+
+  // setup devtools in dev mode
+  if (import.meta.env.DEV || __VUE_PROD_DEVTOOLS__) {
+    import('./devtools').then(({ setupDevtools }) =>
+      setupDevtools(app, router, data)
+    )
   }
 
   return { app, router, data }
@@ -113,25 +111,6 @@ function newRouter(): Router {
 
     return import(/*@vite-ignore*/ pageFilePath)
   }, NotFound)
-}
-
-function handleHMR(router: Router): void {
-  // update route.data on HMR updates of active page
-  if (import.meta.hot) {
-    // hot reload pageData
-    import.meta.hot!.on('vitepress:pageData', (payload) => {
-      if (shouldHotReload(payload)) {
-        router.route.data = payload.pageData
-      }
-    })
-  }
-}
-
-function shouldHotReload(payload: any): boolean {
-  const payloadPath = payload.path.replace(/(\bindex)?\.md$/, '')
-  const locationPath = location.pathname.replace(/(\bindex)?\.html$/, '')
-
-  return payloadPath === locationPath
 }
 
 if (inBrowser) {