+++ src/client/theme-default/components/VPNavBarTitle.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { useData } from 'vitepress'
 import { useSidebar } from '../composables/sidebar'
+import VPImage from './VPImage.vue'
 
 const { site, theme } = useData()
 const { hasSidebar } = useSidebar()
@@ -8,10 +9,12 @@ const { hasSidebar } = useSidebar()
 
 <template>
   <div class="VPNavBarTitle" :class="{ 'has-sidebar': hasSidebar }">
-    <a class="title" href="/">
-      <img v-if="theme.logo" class="logo" :src="theme.logo">
+    <a class="title" :href="site.base">
+      <slot name="nav-bar-title-before" />
+      <VPImage class="logo" :image="theme.logo" />
       <template v-if="theme.siteTitle">{{ theme.siteTitle }}</template>
       <template v-else-if="theme.siteTitle === undefined">{{ site.title }}</template>
+      <slot name="nav-bar-title-after" />
     </a>
   </div>
 </template>
@@ -20,7 +23,6 @@ const { hasSidebar } = useSidebar()
 .VPNavBarTitle {
   flex-shrink: 0;
   border-bottom: 1px solid transparent;
-  transition: background-color 0.5s;
 }
 
 @media (min-width: 960px) {
@@ -53,7 +55,7 @@ const { hasSidebar } = useSidebar()
   }
 }
 
-.logo {
+:deep(.logo) {
   margin-right: 8px;
   height: 24px;
 }