+++ src/node/config.ts
@@ -66,6 +66,13 @@ export interface UserConfig<ThemeConfig = any> {
   mpa?: boolean
 
   /**
+   * Don't fail builds due to dead links.
+   *
+   * @default false
+   */
+  ignoreDeadLinks?: boolean
+
+  /**
    * Build end hook: called when SSG finish.
    * @param siteConfig The resolved configuration.
    */
@@ -80,7 +87,14 @@ export type RawConfigExports<ThemeConfig = any> =
 export interface SiteConfig<ThemeConfig = any>
   extends Pick<
     UserConfig,
-    'markdown' | 'vue' | 'vite' | 'shouldPreload' | 'mpa' | 'lastUpdated' | 'buildEnd'
+    | 'markdown'
+    | 'vue'
+    | 'vite'
+    | 'shouldPreload'
+    | 'mpa'
+    | 'lastUpdated'
+    | 'ignoreDeadLinks'
+    | 'buildEnd'
   > {
   root: string
   srcDir: string
@@ -159,6 +173,7 @@ export async function resolveConfig(
     vite: userConfig.vite,
     shouldPreload: userConfig.shouldPreload,
     mpa: !!userConfig.mpa,
+    ignoreDeadLinks: userConfig.ignoreDeadLinks,
     buildEnd: userConfig.buildEnd
   }
 
@@ -173,14 +188,9 @@ async function resolveUserConfig(
   mode: string
 ): Promise<[UserConfig, string | undefined]> {
   // load user config
-  let configPath
-  for (const ext of supportedConfigExtensions) {
-    const p = resolve(root, `config.${ext}`)
-    if (await fs.pathExists(p)) {
-      configPath = p
-      break
-    }
-  }
+  const configPath = supportedConfigExtensions
+    .map((ext) => resolve(root, `config.${ext}`))
+    .find(fs.pathExistsSync)
 
   const userConfig: RawConfigExports = configPath
     ? ((